fixed wrong server setup
