+++ package.json
@@ -1,40 +1,43 @@
 {
-  "name": "contract-bridger",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "npx tsc",
-    "start": "node src/app.ts",
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/index.js\"",
-    "pretty": "prettier --write \"./**/*.{ts,js,jsx,json}\""
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ContractBridger/contract-bridger.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ContractBridger/contract-bridger/issues"
-  },
-  "homepage": "https://github.com/ContractBridger/contract-bridger#readme",
-  "dependencies": {
-    "axios": "^1.5.1",
-    "commander": "^11.1.0",
-    "dotenv": "^16.3.1",
-    "ethers": "^6.8.0",
-    "express": "^4.18.2",
-    "inquirer": "^9.2.11",
-    "prettier": "^3.0.3",
-    "solc": "^0.8.21"
-  },
-  "devDependencies": {
-    "@types/express": "^4.17.20",
-    "@types/node": "^20.8.7",
-    "concurrently": "^8.2.2",
-    "nodemon": "^3.0.1",
-    "typescript": "^5.2.2"
-  }
+    "name": "contract-bridger",
+    "version": "1.0.0",
+    "description": "",
+    "main": "index.js",
+    "scripts": {
+        "build": "npx tsc",
+        "start": "node dist/src/index.js",
+        "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/src/index.js\"",
+        "pretty": "prettier --write \"./**/*.{ts,js,jsx,json}\""
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/ContractBridger/contract-bridger.git"
+    },
+    "author": "",
+    "license": "ISC",
+    "bugs": {
+        "url": "https://github.com/ContractBridger/contract-bridger/issues"
+    },
+    "homepage": "https://github.com/ContractBridger/contract-bridger#readme",
+    "dependencies": {
+        "axios": "^1.5.1",
+        "commander": "^11.1.0",
+        "cors": "^2.8.5",
+        "dotenv": "^16.3.1",
+        "ethers": "^6.8.0",
+        "express": "^4.18.2",
+        "helmet": "^7.0.0",
+        "inquirer": "^9.2.11",
+        "prettier": "^3.0.3",
+        "solc": "^0.8.21"
+    },
+    "devDependencies": {
+        "@types/cors": "^2.8.15",
+        "@types/express": "^4.17.20",
+        "@types/inquirer": "^9.0.6",
+        "@types/node": "^20.8.7",
+        "concurrently": "^8.2.2",
+        "nodemon": "^3.0.1",
+        "typescript": "^5.2.2"
+    }
 }
