--- src/app.ts
@@ -1,16 +1,0 @@
-import express, { Express, Request, Response } from "express";
-
-import dotenv from "dotenv";
-
-dotenv.config();
-
-const app: Express = express();
-const port = process.env.PORT;
-
-app.get("/", (req: Request, res: Response) => {
-  res.send("");
-});
-
-app.listen(port, () => {
-  console.log(`⚡️[server]: Server is running at http://localhost:${port}`);
-});
