+++ src/index.ts
@@ -1,4 +1,4 @@
-import express from "express";
+import express, { Response } from "express";
 import helmet from "helmet";
 import cors from "cors";
 import routes from "./routes";
@@ -6,11 +6,15 @@ import "dotenv/config";
 
 const app = express();
 
+app.use(helmet());
 app.use(cors());
 app.use(express.json());
+
+app.get("/", (_, res: Response) => {
+    res.status(200).json({ message: "success" });
+});
 app.use(routes);
-app.use(helmet());
 
 app.listen(process.env.PORT || 8080, () => {
-  console.log(`Server running in port ${process.env.PORT || 8080}`);
+    console.log(`Server running in port ${process.env.PORT || 8080}`);
 });
